clean up rebase
