+++ .env
@@ -17,4 +17,4 @@ REACT_APP_VX_DISABLE_BALLOT_BOX_CHECK=FALSE
 REACT_APP_VX_CONVERTER=ms-sems
 REACT_APP_VX_PRECINCT_REPORT_DESTINATION=thermal-sheet-printer
 REACT_APP_VX_DISABLE_BALLOT_BOX_CHECK=FALSE
-REACT_APP_VX_DISABLE_VXMARKSCAN_STATE_MACHINE_POLLING=FALSE
+REACT_APP_VX_SKIP_PAPER_HANDLER_HARDWARE_CHECK=FALSE
