+++ libs/custom-paper-handler/src/driver/index.ts
@@ -6,3 +6,4 @@ export * from './coders';
 export * from './constants';
 export * from './helpers';
 export * from './minimal_web_usb_device';
+export * from './test_utils';
